+++ src/pages/DailyPage.jsx
@@ -1,12 +1,12 @@
-import React, { useEffect, useState } from 'react';
-import MyTable from '../components/MyTable';
-import PageLayout from '../components/PageLayout';
-import MyDialog from '../components/MyDialog';
-import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
-import CustomTableHeading from '../components/CustomTableHeading';
-import useTable from '../hooks/useTable';
-import * as Yup from 'yup';
+import React, { useEffect, useState } from "react";
+import MyTable from "../components/MyTable";
+import PageLayout from "../components/PageLayout";
+import MyDialog from "../components/MyDialog";
+import Box from "@mui/material/Box";
+import Button from "@mui/material/Button";
+import CustomTableHeading from "../components/CustomTableHeading";
+import useTable from "../hooks/useTable";
+import * as Yup from "yup";
 import {
   addDaily,
   addDailyRow,
@@ -14,24 +14,24 @@ import {
   deleteDailyRow,
   editDaily,
   editDailyRow,
-} from '../store/dailyRows';
-import PageHeading from '../components/PageHeading';
-import axios from '../api';
-import { useSelector, useDispatch } from 'react-redux';
-import { getDaily } from '../store/dailyRows';
-import Swal from 'sweetalert2';
-import CircularProgress from '@mui/material/CircularProgress/CircularProgress';
-import { getUsers } from '../store/users';
-const PAGE_TITLE = 'اليومية';
+} from "../store/dailyRows";
+import PageHeading from "../components/PageHeading";
+import axios from "../api";
+import { useSelector, useDispatch } from "react-redux";
+import { getDaily } from "../store/dailyRows";
+import Swal from "sweetalert2";
+import CircularProgress from "@mui/material/CircularProgress/CircularProgress";
+import { getUsers } from "../store/users";
+const PAGE_TITLE = "اليومية";
 
 const DAILY_ROW_INTIAL_VALUE = {
-  accountId: '',
-  id: '',
-  accountName: '',
-  value: '',
-  values: '',
-  billType: '',
-  note: '',
+  accountId: "",
+  id: "",
+  accountName: "",
+  value: "",
+  values: "",
+  billType: "",
+  note: "",
 };
 
 const DailyPage = () => {
@@ -73,8 +73,8 @@ const DailyPage = () => {
       const formattedRow = {
         isDaily: true,
         userId: row.accountId, // its id
-        value: parseInt(row.value, 10),
-        values: parseInt(row.values, 10),
+        value: parseFloat(row.value, 10),
+        values: parseFloat(row.values, 10),
         billType: row.billType,
         note: row.note,
       };
@@ -86,8 +86,8 @@ const DailyPage = () => {
       const formattedRow = {
         isDaily: true,
         userId: editedRow.accountId,
-        value: parseInt(editedRow.value, 10),
-        values: parseInt(editedRow.values, 10),
+        value: parseFloat(editedRow.value, 10),
+        values: parseFloat(editedRow.values, 10),
         billType: editedRow.billType,
         note: editedRow.note,
         id: editedRow.id,
@@ -102,11 +102,11 @@ const DailyPage = () => {
   };
 
   const validationSchema = Yup.object({
-    accountId: Yup.string().required('يرجى إدخال اسم الحساب'),
-    value: Yup.number().required('يرجى إدخال المبلغ'),
-    values: Yup.number().required('يرجى إدخال القيمة'),
+    accountId: Yup.string().required("يرجى إدخال اسم الحساب"),
+    value: Yup.number().required("يرجى إدخال المبلغ"),
+    values: Yup.number().required("يرجى إدخال القيمة"),
     note: Yup.string(),
-    billType: Yup.string().required('يرجى اختيار نوع الفاتورة'),
+    billType: Yup.string().required("يرجى اختيار نوع الفاتورة"),
   });
   const formattedUsers = users?.data?.map((user) => ({
     value: user.name,
@@ -116,67 +116,67 @@ const DailyPage = () => {
 
   const COLUMNS = [
     {
-      id: 'id',
-      label: 'رقم الفاتورة',
+      id: "id",
+      label: "رقم الفاتورة",
       minWidth: 70,
-      align: 'center',
+      align: "center",
       isField: false,
       isLink: true,
       isInvoice: true,
     },
     {
-      id: 'accountId',
-      label: 'ايدي الحساب',
+      id: "accountId",
+      label: "ايدي الحساب",
       minWidth: 80,
-      align: 'center',
+      align: "center",
       isField: true,
       isLink: true,
       options: formattedUsers,
       required: true,
     },
     {
-      id: 'accountName',
-      label: 'اسم الحساب',
+      id: "accountName",
+      label: "اسم الحساب",
       minWidth: 150,
-      align: 'left',
+      align: "left",
     },
     {
-      id: 'value',
-      label: 'المبلغ',
+      id: "value",
+      label: "المبلغ",
       minWidth: 100,
-      align: 'center',
-      format: (value) => value.toLocaleString('en-US'),
+      align: "center",
+      format: (value) => value.toLocaleString("en-US"),
       isField: true,
       required: true,
       isBill: true,
     },
     {
-      id: 'values',
-      label: 'القيمة',
+      id: "values",
+      label: "القيمة",
       minWidth: 50,
-      align: 'center',
-      format: (value) => value.toLocaleString('en-US'),
+      align: "center",
+      format: (value) => value.toLocaleString("en-US"),
       isField: true,
       isBill: true,
     },
     {
-      id: 'billType',
-      label: 'نوع الفاتورة',
-      type: 'select',
+      id: "billType",
+      label: "نوع الفاتورة",
+      type: "select",
       options: [
-        { value: 'ادخال', id: 'ادخال' },
-        { value: 'صادر', id: 'صادر' },
-        { value: 'مصروف', id: 'مصروف' },
-        { value: 'سحوبات', id: 'سحوبات' },
+        { value: "ادخال", id: "ادخال" },
+        { value: "صادر", id: "صادر" },
+        { value: "مصروف", id: "مصروف" },
+        { value: "سحوبات", id: "سحوبات" },
       ],
       required: true,
       isField: true,
     },
     {
-      id: 'note',
-      label: 'ملاحظة',
+      id: "note",
+      label: "ملاحظة",
       minWidth: 170,
-      align: 'left',
+      align: "left",
       isField: true,
     },
   ];
@@ -185,7 +185,7 @@ const DailyPage = () => {
       <PageLayout>
         <PageHeading title="صفحة اليومية" />
         <Box>
-          <CustomTableHeading rows={{ ...yesterday, title: 'الاساس' }} />
+          <CustomTableHeading rows={{ ...yesterday, title: "الاساس" }} />
           <MyTable
             COLUMNS={COLUMNS}
             deletedLabel={`فاتورة ${PAGE_TITLE}`}
@@ -193,19 +193,19 @@ const DailyPage = () => {
             rows={formattedBills}
             dispatchers={dispatchers}
           />
-          <CustomTableHeading rows={{ ...today, title: 'المدور' }} />
+          <CustomTableHeading rows={{ ...today, title: "المدور" }} />
           <Button
             variant="contained"
             sx={{
-              height: '60px',
-              margin: '2rem auto',
-              display: 'block',
-              width: '50%',
+              height: "60px",
+              margin: "2rem auto",
+              display: "block",
+              width: "50%",
             }}
             color="error"
             onClick={handleOpenAddDialog}
           >
-            {'اضافة فاتورة جديدة'}
+            {"اضافة فاتورة جديدة"}
           </Button>
         </Box>
         <MyDialog
