+++ src/pages/PartnersPage/PartnerPage/index.jsx
@@ -1,61 +1,61 @@
-import React, { useEffect, useState } from 'react';
-import MyTable from '../../../components/MyTable';
-import PageLayout from '../../../components/PageLayout';
-import MyDialog from '../../../components/MyDialog';
-import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
-import Typography from '@mui/material/Typography';
-import useTable from '../../../hooks/useTable';
-import { useSelector } from 'react-redux';
-import * as Yup from 'yup';
+import React, { useEffect, useState } from "react";
+import MyTable from "../../../components/MyTable";
+import PageLayout from "../../../components/PageLayout";
+import MyDialog from "../../../components/MyDialog";
+import Box from "@mui/material/Box";
+import Button from "@mui/material/Button";
+import Typography from "@mui/material/Typography";
+import useTable from "../../../hooks/useTable";
+import { useSelector } from "react-redux";
+import * as Yup from "yup";
 import {
   addCustomer,
   deleteCustomer,
   editCustomer,
   getUser,
-} from '../../../store/users';
-import { useDispatch } from 'react-redux';
-import PageHeading from '../../../components/PageHeading';
-import { useNavigate, useParams } from 'react-router-dom';
-import userImage from '../../../assets/userImage.png';
-import { deleteDaily, editDaily, getDaily } from '../../../store//dailyRows';
-const PAGE_TITLE = 'الزبائن';
+} from "../../../store/users";
+import { useDispatch } from "react-redux";
+import PageHeading from "../../../components/PageHeading";
+import { useNavigate, useParams } from "react-router-dom";
+import userImage from "../../../assets/userImage.png";
+import { deleteDaily, editDaily, getDaily } from "../../../store//dailyRows";
+const PAGE_TITLE = "الزبائن";
 const COLUMNS = [
   {
-    id: 'id',
-    label: 'رقم الفاتورة',
+    id: "id",
+    label: "رقم الفاتورة",
     minWidth: 80,
-    align: 'center',
+    align: "center",
     isField: false,
     // isInvoice: true,
     // urlPath: 'i  nvoices',
   },
   {
-    id: 'value',
-    label: 'المبلغ',
+    id: "value",
+    label: "المبلغ",
     minWidth: 100,
-    align: 'center',
+    align: "center",
     isField: true,
     required: true,
   },
   {
-    id: 'values',
-    label: 'القيمة',
+    id: "values",
+    label: "القيمة",
     minWidth: 80,
-    align: 'center',
+    align: "center",
     isField: true,
     required: true,
   },
   {
-    id: 'billType',
-    label: 'النوع',
+    id: "billType",
+    label: "النوع",
     minWidth: 150,
-    align: 'center',
+    align: "center",
     isField: false,
     required: true,
     options: [
-      { value: 'ادخال', id: '1' },
-      { value: 'صادر', id: '2' },
+      { value: "ادخال", id: "1" },
+      { value: "صادر", id: "2" },
     ],
   },
   // {
@@ -66,29 +66,29 @@ const COLUMNS = [
   //   isField: true,
   // },
   {
-    id: 'note',
-    label: 'ملاحظة',
+    id: "note",
+    label: "ملاحظة",
     minWidth: 170,
-    align: 'left',
+    align: "left",
     isField: true,
   },
   {
-    id: 'createdAt',
-    label: 'التاريخ',
+    id: "createdAt",
+    label: "التاريخ",
     minWidth: 200,
-    align: 'center',
+    align: "center",
   },
 ];
 const DAILY_ROW_INTIAL_VALUE = {
   value: 0,
   values: 0,
-  billType: 'ادخال',
-  note: '',
+  billType: "ادخال",
+  note: "",
 };
 const validationSchema = Yup.object().shape({
-  value: Yup.string().required('المبلغ مطلوب'),
-  values: Yup.string().required('القيمة مطلوبة'),
-  billType: Yup.string().required('نوع الفاتورة مطلوب'),
+  value: Yup.string().required("المبلغ مطلوب"),
+  values: Yup.string().required("القيمة مطلوبة"),
+  billType: Yup.string().required("نوع الفاتورة مطلوب"),
   note: Yup.string(),
 });
 const index = () => {
@@ -96,21 +96,21 @@ const index = () => {
   const navigate = useNavigate();
   const user = useSelector((state) => state.users.user);
   const pageFields = [
-    { fieldName: 'رقم الحساب', value: user.id || 'لايوجد' },
-    { fieldName: 'النوع', value: user.userType || 'لايوجد' },
+    { fieldName: "رقم الحساب", value: user.id || "لايوجد" },
+    { fieldName: "النوع", value: user.userType || "لايوجد" },
 
-    { fieldName: 'رقم الهاتف', value: user.mobileNumber || 'لايوجد' },
+    { fieldName: "رقم الهاتف", value: user.mobileNumber || "لايوجد" },
 
-    { fieldName: 'العنوان', value: user.address || 'لايوجد' },
-    { fieldName: 'ملاحظة', value: user.note || 'لايوجد' },
+    { fieldName: "العنوان", value: user.address || "لايوجد" },
+    { fieldName: "ملاحظة", value: user.note || "لايوجد" },
     {
-      fieldName: 'صافي الحساب',
-      value: user.accountBalance || 'لايوجد',
+      fieldName: "صافي الحساب",
+      value: user.accountBalance || "لايوجد",
       isMoney: true,
     },
     {
-      fieldName: 'قيمة صافي الحساب',
-      value: user.accountBalanceValues || 'لايوجد',
+      fieldName: "قيمة صافي الحساب",
+      value: user.accountBalanceValues || "لايوجد",
       isMoney: true,
     },
   ];
@@ -122,8 +122,8 @@ const index = () => {
       dispatch(
         editDaily({
           ...editedDaily,
-          value: parseInt(editedDaily.value),
-          values: parseInt(editedDaily.values),
+          value: parseFloat(editedDaily.value),
+          values: parseFloat(editedDaily.values),
           userId: editedDaily.UserId,
         })
       ).then(() => dispatch(getUser(params.id)));
@@ -144,20 +144,20 @@ const index = () => {
   return (
     <PageLayout title="صفحة زبون">
       <Box>
-        <Box sx={{ width: '100px', overflow: 'hidden', margin: 'auto', mb: 3 }}>
+        <Box sx={{ width: "100px", overflow: "hidden", margin: "auto", mb: 3 }}>
           <img
             src={userImage}
             alt="user image"
             width="100%"
-            style={{ objectFit: 'contain' }}
+            style={{ objectFit: "contain" }}
           />
         </Box>
         <Box sx={{ mb: 6 }}>
           <Typography
             sx={{
-              textAlign: 'center',
-              fontWeight: 'bold',
-              fontSize: '2rem',
+              textAlign: "center",
+              fontWeight: "bold",
+              fontSize: "2rem",
               mb: 4,
             }}
           >
@@ -167,29 +167,29 @@ const index = () => {
             <Box sx={{ mb: 3 }} key={pf.fieldName}>
               <Typography
                 sx={{
-                  display: 'inline-block',
+                  display: "inline-block",
                   mr: 4,
-                  fontWeight: 'bold',
-                  fontSize: '1.3rem',
-                  minWidth: '120px',
+                  fontWeight: "bold",
+                  fontSize: "1.3rem",
+                  minWidth: "120px",
                 }}
               >
                 {`${pf.fieldName}:`}
               </Typography>
               <Typography
                 sx={{
-                  display: 'inline-block',
-                  direction: pf.isMoney && 'rtl',
-                  textAlign: pf.isMoney && 'center',
+                  display: "inline-block",
+                  direction: pf.isMoney && "rtl",
+                  textAlign: pf.isMoney && "center",
 
                   color: pf.isMoney
                     ? pf.value >= 0
-                      ? 'green'
-                      : 'tomato'
-                    : 'inherit',
-                  fontSize: '1.3rem',
-                  fontWeight: pf.isMoney && 'bold',
-                  minWidth: '120px',
+                      ? "green"
+                      : "tomato"
+                    : "inherit",
+                  fontSize: "1.3rem",
+                  fontWeight: pf.isMoney && "bold",
+                  minWidth: "120px",
                 }}
               >
                 {`${pf.value}`}
